Dashboard: Moved data folder into components
