+++ Frontend/src/components/DashboardComponents/RevenueChart.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Line } from "react-chartjs-2";
 import { Chart as ChartJS, defaults } from "chart.js/auto";
-import  SourceData  from "../../data/SourceData.json";
+import  SourceData from "../data/SourceData.json";
 
 defaults.maintainAspectRatio = false;
 defaults.responsive = true;
